removes duplciate config service and fixes tests
